* commit '07298fb35c4adca8bfb0e96e0c7d508ef87c0356':
  Add help flag to avoid error when running CLI with yaml config
  Removing unecessary ignore rule
  added seperate tests for docker (pip) and docker (conda)
  Add conda dockerfile and logic for pip vs conda docker build
  feat: start to build dockerfile for conda

+++ hooks/post_gen_project.py
@@ -3,7 +3,7 @@ import shutil
 import sys
 from glob import glob
 from pathlib import Path
-from typing import List
+from typing import Iterable, Optional
 
 module_dir = 'src/{{ cookiecutter.module_name }}'
 files_pip = [
@@ -14,11 +14,19 @@ files_conda = [
     'environment.yml',
     'environment-dev.yml',
 ]
-files_docker = [
-    'Dockerfile',
+files_docker_aux = [
     'docker-compose.yml',
     '.dockerignore',
 ]
+
+files_dockerfile_pip = [
+    'Dockerfile__pip',
+]
+
+files_dockerfile_conda = [
+    'Dockerfile__conda',
+]
+
 files_cli = [
     f'{module_dir}/main__cli.py'
 ]
@@ -57,7 +65,17 @@ def handle_notebooks():
 def handle_docker():
     use_docker = '{{ cookiecutter.use_docker }}'
     if use_docker == 'no':
-        _delete_files(files_docker)
+        _delete_files(files_docker_aux + files_dockerfile_pip + files_dockerfile_conda)
+    else:
+        package_manager = '{{ cookiecutter.package_manager }}'
+        if package_manager == "conda":
+            _delete_files(files_dockerfile_pip)
+            _rename_files("Dockerfile__conda", "__conda", "")
+        elif package_manager == "pip":
+            _delete_files(files_dockerfile_conda)
+            _rename_files("Dockerfile__pip", "__pip", "")
+
+    
 
 
 def handle_cli():
@@ -110,6 +128,7 @@ def handle_editor_settings():
         sys.exit(1)
 
 
+
 def print_success():
     full_name = '{{ cookiecutter.full_name }}'
     print(f"Hey {full_name}! Your project was successfully created at {os.getcwd()}. "
@@ -122,16 +141,17 @@ def _rename_files(file_pattern, old, new):
         path.rename(path.with_name(path.name.replace(old, new)))
 
 
-def _delete_files(files):
+def _delete_files(files: Iterable[str], exclude: Optional[str] = None):
     try:
         for file in files:
-            os.remove(file)
+            if file != exclude:
+                os.remove(file)
     except OSError as e:
         print(f"Error: failed to remove files - {e}")
         sys.exit(1)
 
 
-def _delete_folders(folders: List[str], exclude: str = None):
+def _delete_folders(folders: Iterable[str], exclude: Optional[str] = None):
     for folder in folders:
         if folder != exclude:
             shutil.rmtree(folder, ignore_errors=True)
